refactor: 모바일 tab scroll 로직 분리
